refactor(entities): add ShipRotator
